# backend/app/src/routes/swagger.json
{
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "ExamJam Web API",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/exams/{examId}": {
			"get": {
				"operationId": "GetFile",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Exam"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "examId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}